refactor(core): relocate playbook engine into core/domain (hexagonal P4 sweep)
Physical move of the playbook evaluation engine (engine/evaluator/loader/
conditions/context/presentation/schema_registry/templates/verdict) from
regis/playbook/ into regis/core/domain/playbook/. Cosmetic placement only —
the layering was already enforced by import-linter regardless of location.

The package only depends on regis.core.domain (rules) and its own submodules,
so it lands cleanly in the domain layer; the rules<->playbook cycle stays
intra-domain (unconstrained within a single layer). Schema resources stay
anchored to regis.schemas (no package-data change). The standalone-mypy
follow_imports override is retargeted to the new module path.
