+++ regis/core/domain/playbook/schema_registry.py
