+++ tests/test_coverage_engine.py
@@ -1,4 +1,4 @@
-from regis.playbook.engine import (
+from regis.core.domain.playbook.engine import (
     MissingDataTracker,
     _format_date,
     _format_datetime,
